requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
dnspython==1.16.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
fonttools==4.28.5
gunicorn==20.1.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mysql-connector-python==8.0.26
numpy==1.21.2
outcome==1.1.0
packaging==21.3
pandas==1.3.3
pbu==0.7.2
Pillow==8.4.0
protobuf==3.18.1
pycparser==2.20
pymongo==4.1.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selenium==4.0.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
threadpoolctl==3.0.0
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
